refactor(reporters): split `console_reporter` statistical+coverage into mixins
`ConsoleReporter` had grown to 1303 lines because three responsibilities lived
on the same class: core single-run output (summary/detail/timeline), behavior
coverage reporting, and statistical/variance reporting (pass@k, flakiness
panels, individual-runs tables, multi-run comparison). The first is what the
class is named for; the other two are independent feature sets that touched
nothing else on the class besides `self.console`.

Extracted into mixin modules following the same pattern as
`SecurityChecksMixin` from #202:

- `_console_coverage.py` — `CoverageReporterMixin` (`print_coverage_report`)
- `_console_statistical.py` — `StatisticalReporterMixin`
  (`print_statistical_summary`, `_print_statistics_table`,
  `_get_variance_indicator`, `_print_flakiness_panel`,
  `_print_individual_runs_table`, `print_statistical_comparison`)

`ConsoleReporter` is now 842 lines and inherits from both mixins, so every
public method (`print_summary`, `print_coverage_report`,
`print_statistical_summary`, …) is still resolvable on a plain
`ConsoleReporter()` instance — no caller changes needed.

Pure refactor: ruff/mypy clean, 45 reporter tests pass.
