+++ evalview/reporters/console_reporter.py
@@ -10,15 +10,13 @@ from rich.text import Text
 from evalview.core.types import (
     EvaluationResult,
     StepTrace,
-    TestCase,
-    StatisticalEvaluationResult,
-    StatisticalMetrics,
-    FlakinessScore,
     ReasonCode,
 )
+from evalview.reporters._console_coverage import CoverageReporterMixin
+from evalview.reporters._console_statistical import StatisticalReporterMixin
 
 
-class ConsoleReporter:
+class ConsoleReporter(StatisticalReporterMixin, CoverageReporterMixin):
     """Generates formatted console output for evaluation results."""
 
     def __init__(self):
@@ -842,462 +840,3 @@ class ConsoleReporter:
 
         return deltas
 
-    def print_coverage_report(
-        self,
-        test_cases: List[TestCase],
-        results: List[EvaluationResult],
-        suite_name: Optional[str] = None,
-    ) -> None:
-        """Print a behavior coverage report.
-
-        Shows coverage across:
-        - Tasks: Scenarios tested
-        - Tools: Agent tools exercised
-        - Paths: Multi-step workflows
-        - Eval dimensions: Correctness, safety, cost, latency checks
-
-        Args:
-            test_cases: List of test case definitions.
-            results: List of evaluation results.
-            suite_name: Optional name for the test suite.
-        """
-        if not test_cases:
-            self.console.print("[yellow]No test cases to analyze[/yellow]")
-            return
-
-        # Header
-        self.console.print()
-        self.console.print("[bold]━━━ Behavior Coverage ━━━[/bold]")
-
-        if suite_name:
-            self.console.print(f"[dim]Suite:[/dim] {suite_name}")
-        self.console.print()
-
-        # 1. Tasks Coverage
-        total_tasks = len(test_cases)
-        executed_tasks = len(results)
-        passed_tasks = sum(1 for r in results if r.passed)
-        task_pct = (executed_tasks / total_tasks * 100) if total_tasks > 0 else 0
-
-        task_color = "green" if task_pct == 100 else "yellow" if task_pct >= 50 else "red"
-        self.console.print(f"[bold]Tasks:[/bold]      [{task_color}]{executed_tasks}/{total_tasks} scenarios ({task_pct:.0f}%)[/{task_color}]")
-        if passed_tasks < executed_tasks:
-            self.console.print(f"            [dim]({passed_tasks} passing, {executed_tasks - passed_tasks} failing)[/dim]")
-
-        # 2. Tools Coverage
-        # Collect all expected tools from test cases
-        expected_tools = set()
-        for tc in test_cases:
-            if tc.expected.tools:
-                expected_tools.update(tc.expected.tools)
-            if tc.expected.tool_sequence:
-                expected_tools.update(tc.expected.tool_sequence)
-            if tc.expected.sequence:
-                expected_tools.update(tc.expected.sequence)
-
-        # Collect all actually called tools from results
-        exercised_tools = set()
-        for result in results:
-            if result.trace.steps:
-                for step in result.trace.steps:
-                    if step.tool_name:
-                        exercised_tools.add(step.tool_name)
-
-        # Also add tools from evaluations
-        for result in results:
-            if result.evaluations.tool_accuracy.correct:
-                exercised_tools.update(result.evaluations.tool_accuracy.correct)
-
-        # Calculate tool coverage
-        if expected_tools:
-            tools_covered = expected_tools & exercised_tools
-            tool_pct = (len(tools_covered) / len(expected_tools) * 100) if expected_tools else 0
-            tool_color = "green" if tool_pct == 100 else "yellow" if tool_pct >= 50 else "red"
-            self.console.print(f"[bold]Tools:[/bold]      [{tool_color}]{len(tools_covered)}/{len(expected_tools)} exercised ({tool_pct:.0f}%)[/{tool_color}]")
-
-            # Show missing tools
-            missing_tools = expected_tools - exercised_tools
-            if missing_tools:
-                self.console.print(f"            [dim]missing: {', '.join(sorted(missing_tools))}[/dim]")
-        else:
-            self.console.print("[bold]Tools:[/bold]      [dim]no tool expectations defined[/dim]")
-
-        # 3. Paths Coverage (multi-step workflows)
-        # Count tests with sequence requirements
-        sequence_tests = [tc for tc in test_cases if tc.expected.tool_sequence or tc.expected.sequence]
-        total_paths = len(sequence_tests)
-
-        if total_paths > 0:
-            # Check which sequence tests passed
-            sequence_test_names = {tc.name for tc in sequence_tests}
-            sequence_results = [r for r in results if r.test_case in sequence_test_names]
-            paths_passed = sum(1 for r in sequence_results if r.evaluations.sequence_correctness.correct)
-
-            path_pct = (paths_passed / total_paths * 100) if total_paths > 0 else 0
-            path_color = "green" if path_pct == 100 else "yellow" if path_pct >= 50 else "red"
-            self.console.print(f"[bold]Paths:[/bold]      [{path_color}]{paths_passed}/{total_paths} multi-step workflows ({path_pct:.0f}%)[/{path_color}]")
-        else:
-            self.console.print("[bold]Paths:[/bold]      [dim]no sequence tests defined[/dim]")
-
-        # 4. Eval Dimensions
-        self.console.print("[bold]Dimensions:[/bold]")
-
-        # Check which dimensions are being tested
-        has_tool_check = any(tc.expected.tools or tc.expected.tool_sequence for tc in test_cases)
-        has_output_check = any(tc.expected.output for tc in test_cases)
-        has_cost_check = any(tc.thresholds.max_cost is not None for tc in test_cases)
-        has_latency_check = any(tc.thresholds.max_latency is not None for tc in test_cases)
-        has_hallucination_check = any(
-            tc.expected.hallucination is not None or (tc.checks is None or tc.checks.hallucination)
-            for tc in test_cases
-        )
-        has_safety_check = any(
-            tc.expected.safety is not None or (tc.checks is None or tc.checks.safety)
-            for tc in test_cases
-        )
-
-        dimensions = []
-        if has_tool_check:
-            # Check if tool checks pass
-            tool_pass = all(r.evaluations.tool_accuracy.accuracy == 1.0 for r in results) if results else False
-            dimensions.append(("correctness", tool_pass))
-        if has_output_check:
-            output_pass = all(r.evaluations.output_quality.score >= 70 for r in results) if results else False
-            dimensions.append(("output", output_pass))
-        if has_cost_check:
-            cost_pass = all(r.evaluations.cost.passed for r in results) if results else False
-            dimensions.append(("cost", cost_pass))
-        if has_latency_check:
-            latency_pass = all(r.evaluations.latency.passed for r in results) if results else False
-            dimensions.append(("latency", latency_pass))
-        if has_hallucination_check:
-            hall_pass = all(
-                r.evaluations.hallucination is None or r.evaluations.hallucination.passed
-                for r in results
-            ) if results else False
-            dimensions.append(("hallucination", hall_pass))
-        if has_safety_check:
-            safety_pass = all(
-                r.evaluations.safety is None or r.evaluations.safety.passed
-                for r in results
-            ) if results else False
-            dimensions.append(("safety", safety_pass))
-
-        if dimensions:
-            dim_strs = []
-            for name, passed in dimensions:
-                icon = "[green]✓[/green]" if passed else "[red]✗[/red]"
-                dim_strs.append(f"{name} {icon}")
-            self.console.print(f"            {', '.join(dim_strs)}")
-        else:
-            self.console.print("            [dim]no thresholds configured[/dim]")
-
-        # Overall coverage score
-        self.console.print()
-        coverage_scores = []
-        if total_tasks > 0:
-            coverage_scores.append(task_pct)
-        if expected_tools:
-            coverage_scores.append(tool_pct)
-        if total_paths > 0:
-            coverage_scores.append(path_pct)
-
-        if coverage_scores:
-            overall_coverage = sum(coverage_scores) / len(coverage_scores)
-            cov_color = "green" if overall_coverage >= 80 else "yellow" if overall_coverage >= 50 else "red"
-            self.console.print(f"[bold]Overall:[/bold]    [{cov_color}]{overall_coverage:.0f}% behavior coverage[/{cov_color}]")
-
-        self.console.print()
-
-    # --- Statistical/Variance Reporting ---
-
-    def print_statistical_summary(
-        self,
-        result: StatisticalEvaluationResult,
-        show_individual_runs: bool = False,
-    ) -> None:
-        """Print a comprehensive statistical evaluation summary.
-
-        Args:
-            result: Statistical evaluation result.
-            show_individual_runs: Whether to show details of each individual run.
-        """
-        # Header with pass/fail status
-        status_icon = "✅" if result.passed else "❌"
-        status_color = "green" if result.passed else "red"
-        status_text = "PASSED" if result.passed else "FAILED"
-
-        self.console.print()
-        self.console.print(
-            f"[bold {status_color}]{status_icon} Statistical Evaluation: {result.test_case}[/bold {status_color}]"
-        )
-        self.console.print(f"[bold {status_color}]{status_text}[/bold {status_color}]")
-        self.console.print()
-
-        # Run summary panel
-        pass_rate_color = "green" if result.pass_rate >= 0.8 else "yellow" if result.pass_rate >= 0.5 else "red"
-
-        # pass@k interpretation
-        pass_at_k_color = "green" if result.pass_at_k >= 0.95 else "yellow" if result.pass_at_k >= 0.8 else "red"
-        pass_at_k_meaning = "usually finds a solution" if result.pass_at_k >= 0.8 else "inconsistent"
-
-        # pass^k interpretation
-        pass_power_k_color = "green" if result.pass_power_k >= 0.5 else "yellow" if result.pass_power_k >= 0.2 else "red"
-        pass_power_k_meaning = "reliable" if result.pass_power_k >= 0.5 else "needs improvement" if result.pass_power_k >= 0.2 else "unreliable"
-
-        run_summary = (
-            f"  [bold]Total Runs:[/bold]     {result.total_runs}\n"
-            f"  [bold]Passed:[/bold]         [green]{result.successful_runs}[/green]\n"
-            f"  [bold]Failed:[/bold]         [red]{result.failed_runs}[/red]\n"
-            f"  [bold]Pass Rate:[/bold]      [{pass_rate_color}]{result.pass_rate:.1%}[/{pass_rate_color}] "
-            f"(required: {result.required_pass_rate:.1%})\n"
-            f"\n"
-            f"  [bold]Reliability Metrics:[/bold]\n"
-            f"  [bold]pass@{result.total_runs}:[/bold]       [{pass_at_k_color}]{result.pass_at_k:.1%}[/{pass_at_k_color}] "
-            f"[dim]({pass_at_k_meaning})[/dim]\n"
-            f"  [bold]pass^{result.total_runs}:[/bold]       [{pass_power_k_color}]{result.pass_power_k:.1%}[/{pass_power_k_color}] "
-            f"[dim]({pass_power_k_meaning})[/dim]"
-        )
-        self.console.print(Panel(run_summary, title="[bold]Run Summary[/bold]", border_style="cyan"))
-
-        # Score statistics table
-        self._print_statistics_table(result.score_stats, "Score Statistics", unit="pts")
-
-        # Cost statistics (if available)
-        if result.cost_stats:
-            self._print_statistics_table(result.cost_stats, "Cost Statistics", unit="$", precision=4)
-
-        # Latency statistics (if available)
-        if result.latency_stats:
-            self._print_statistics_table(result.latency_stats, "Latency Statistics", unit="ms", precision=0)
-
-        # Flakiness assessment
-        self._print_flakiness_panel(result.flakiness)
-
-        # Failure reasons (if any)
-        if result.failure_reasons:
-            self.console.print()
-            self.console.print("[bold red]Failure Reasons:[/bold red]")
-            for reason in result.failure_reasons:
-                self.console.print(f"  [yellow]• {reason}[/yellow]")
-
-        # Individual run details (optional)
-        if show_individual_runs and result.individual_results:
-            self._print_individual_runs_table(result.individual_results)
-
-        # Configuration summary
-        config = result.variance_config
-        self.console.print()
-        self.console.print("[dim]Configuration:[/dim]")
-        self.console.print(f"  [dim]runs: {config.runs}, pass_rate: {config.pass_rate}, confidence: {config.confidence_level}[/dim]")
-
-    def _print_statistics_table(
-        self,
-        stats: StatisticalMetrics,
-        title: str,
-        unit: str = "",
-        precision: int = 2,
-    ) -> None:
-        """Print a formatted statistics table.
-
-        Args:
-            stats: Statistical metrics to render.
-            title: Table title.
-            unit: Unit suffix to display (e.g. "$", "ms", "pts").
-            precision: Decimal precision used for formatted values.
-        """
-        self.console.print()
-
-        table = Table(title=title, show_header=True, header_style="bold cyan")
-        table.add_column("Metric", style="cyan")
-        table.add_column("Value", justify="right")
-        table.add_column("", width=30)  # Visual indicator column
-
-        def fmt(val: float) -> str:
-            if unit == "$":
-                return f"${val:.{precision}f}"
-            elif unit == "ms":
-                return f"{val:.{precision}f}ms"
-            else:
-                return f"{val:.{precision}f}{unit}"
-
-        # Mean with CI
-        ci_str = f"[{stats.confidence_interval_lower:.{precision}f}, {stats.confidence_interval_upper:.{precision}f}]"
-        ci_pct = int(stats.confidence_level * 100)
-        table.add_row("Mean", fmt(stats.mean), f"[dim]{ci_pct}% CI: {ci_str}[/dim]")
-
-        # Std Dev and Variance
-        table.add_row("Std Dev", fmt(stats.std_dev), self._get_variance_indicator(stats.std_dev, stats.mean))
-        table.add_row("Variance", f"{stats.variance:.{precision}f}", "")
-
-        # Min/Max range
-        range_val = stats.max_value - stats.min_value
-        table.add_row("Min", fmt(stats.min_value), "")
-        table.add_row("Max", fmt(stats.max_value), f"[dim]range: {range_val:.{precision}f}[/dim]")
-
-        # Percentiles
-        table.add_row("Median (P50)", fmt(stats.median), "")
-        table.add_row("P25", fmt(stats.percentile_25), "")
-        table.add_row("P75", fmt(stats.percentile_75), "")
-        table.add_row("P95", fmt(stats.percentile_95), "")
-
-        self.console.print(table)
-
-    def _get_variance_indicator(self, std_dev: float, mean: float) -> str:
-        """Get a visual indicator for variance level.
-
-        Args:
-            std_dev: Standard deviation for the distribution.
-            mean: Mean value for the distribution.
-
-        Returns:
-            A Rich markup string describing the variance level.
-        """
-        if mean == 0:
-            return ""
-
-        cv = (std_dev / mean) * 100  # Coefficient of variation
-        if cv < 5:
-            return "[green]▁▁▁▁▁ Low variance[/green]"
-        elif cv < 10:
-            return "[green]▂▂▁▁▁ Low variance[/green]"
-        elif cv < 20:
-            return "[yellow]▄▄▂▁▁ Moderate variance[/yellow]"
-        elif cv < 30:
-            return "[yellow]▆▆▄▂▁ High variance[/yellow]"
-        else:
-            return "[red]█████ Very high variance[/red]"
-
-    def _print_flakiness_panel(self, flakiness: FlakinessScore) -> None:
-        """Print the flakiness assessment panel.
-
-        Args:
-            flakiness: Flakiness score and associated metadata.
-        """
-        self.console.print()
-
-        # Color based on category
-        category_colors = {
-            "stable": "green",
-            "low_variance": "green",
-            "moderate_variance": "yellow",
-            "high_variance": "yellow",
-            "flaky": "red",
-        }
-        color = category_colors.get(flakiness.category, "white")
-
-        # Visual flakiness bar
-        filled = int(flakiness.score * 10)
-        bar = "█" * filled + "░" * (10 - filled)
-
-        content = (
-            f"  [bold]Flakiness Score:[/bold] [{color}]{flakiness.score:.2f}[/{color}] [{color}]{bar}[/{color}]\n"
-            f"  [bold]Category:[/bold]        [{color}]{flakiness.category}[/{color}]\n"
-            f"  [bold]Pass Rate:[/bold]       {flakiness.pass_rate:.1%}\n"
-            f"  [bold]Score CV:[/bold]        {flakiness.score_coefficient_of_variation:.1f}%"
-        )
-
-        if flakiness.contributing_factors and flakiness.contributing_factors != ["none"]:
-            content += "\n\n  [bold]Contributing Factors:[/bold]"
-            for factor in flakiness.contributing_factors:
-                content += f"\n    [dim]• {factor}[/dim]"
-
-        self.console.print(Panel(content, title="[bold]Flakiness Assessment[/bold]", border_style=color))
-
-    def _print_individual_runs_table(self, results: List[EvaluationResult]) -> None:
-        """Print a table showing individual run results.
-
-        Args:
-            results: Per-run evaluation results to render.
-        """
-        self.console.print()
-
-        table = Table(title="Individual Runs", show_header=True, header_style="bold")
-        table.add_column("#", style="dim", width=4)
-        table.add_column("Status", justify="center", width=8)
-        table.add_column("Score", justify="right")
-        table.add_column("Cost", justify="right")
-        table.add_column("Latency", justify="right")
-        table.add_column("Tool Acc", justify="right")
-
-        for i, result in enumerate(results, 1):
-            status = "[green]✓ Pass[/green]" if result.passed else "[red]✗ Fail[/red]"
-            score_color = "green" if result.score >= 80 else "yellow" if result.score >= 60 else "red"
-
-            table.add_row(
-                str(i),
-                status,
-                f"[{score_color}]{result.score:.1f}[/{score_color}]",
-                f"${result.trace.metrics.total_cost:.4f}",
-                f"{result.trace.metrics.total_latency:.0f}ms",
-                f"{result.evaluations.tool_accuracy.accuracy * 100:.0f}%",
-            )
-
-        self.console.print(table)
-
-    def print_statistical_comparison(
-        self,
-        results: List[StatisticalEvaluationResult],
-    ) -> None:
-        """Print a comparison table of multiple statistical evaluations.
-
-        Args:
-            results: Statistical evaluation results to compare.
-        """
-        if not results:
-            self.console.print("[yellow]No results to compare[/yellow]")
-            return
-
-        self.console.print()
-        self.console.print("[bold]━━━ Statistical Comparison ━━━[/bold]")
-        self.console.print()
-
-        table = Table(show_header=True, header_style="bold cyan")
-        table.add_column("Test Case", style="cyan")
-        table.add_column("Status", justify="center")
-        table.add_column("Pass Rate", justify="right")
-        table.add_column("pass@k", justify="right")
-        table.add_column("pass^k", justify="right")
-        table.add_column("Mean Score", justify="right")
-        table.add_column("Flakiness", justify="center")
-
-        for result in results:
-            status = "[green]✓[/green]" if result.passed else "[red]✗[/red]"
-            pass_color = "green" if result.pass_rate >= 0.8 else "yellow" if result.pass_rate >= 0.5 else "red"
-            score_color = "green" if result.score_stats.mean >= 80 else "yellow" if result.score_stats.mean >= 60 else "red"
-
-            # pass@k coloring (high is good - "will it work eventually?")
-            pass_at_k_color = "green" if result.pass_at_k >= 0.95 else "yellow" if result.pass_at_k >= 0.8 else "red"
-
-            # pass^k coloring (reliability - "will it work every time?")
-            pass_power_k_color = "green" if result.pass_power_k >= 0.5 else "yellow" if result.pass_power_k >= 0.2 else "red"
-
-            flakiness_icons = {
-                "stable": "[green]●[/green]",
-                "low_variance": "[green]◐[/green]",
-                "moderate_variance": "[yellow]◑[/yellow]",
-                "high_variance": "[yellow]○[/yellow]",
-                "flaky": "[red]◌[/red]",
-            }
-            flakiness_icon = flakiness_icons.get(result.flakiness.category, "?")
-
-            table.add_row(
-                result.test_case,
-                status,
-                f"[{pass_color}]{result.pass_rate:.1%}[/{pass_color}]",
-                f"[{pass_at_k_color}]{result.pass_at_k:.1%}[/{pass_at_k_color}]",
-                f"[{pass_power_k_color}]{result.pass_power_k:.1%}[/{pass_power_k_color}]",
-                f"[{score_color}]{result.score_stats.mean:.1f}[/{score_color}]",
-                f"{flakiness_icon} {result.flakiness.category}",
-            )
-
-        self.console.print(table)
-
-        # Summary
-        total = len(results)
-        passed = sum(1 for r in results if r.passed)
-        flaky_count = sum(1 for r in results if r.flakiness.category in ("high_variance", "flaky"))
-
-        self.console.print()
-        self.console.print(f"[bold]Summary:[/bold] {passed}/{total} passed, {flaky_count} flaky tests")
-        self.console.print()
